+++ requirements.txt
@@ -1,0 +1,9 @@
+fire
+numpy
+retry
+tqdm
+torch
+torchvision
+pillow
+torch_optimizer
+contrastive_learner>=0.0.8
