+++ stylegan2_pytorch/dataset/Dataset.py
@@ -1,0 +1,36 @@
+from functools import partial
+from pathlib import Path
+
+from PIL import Image
+from torch.utils import data
+from torchvision import transforms
+
+from misc import *
+
+class Dataset(data.Dataset):
+    def __init__(self, folder, image_size, transparent = False):
+        super().__init__()
+        self.folder = folder
+        self.image_size = image_size
+        self.paths = [p for ext in EXTS for p in Path(f'{folder}').glob(f'**/*.{ext}')]
+
+        convert_image_fn = convert_transparent_to_rgb if not transparent else convert_rgb_to_transparent
+        num_channels = 3 if not transparent else 4
+
+        self.transform = transforms.Compose([
+            transforms.Lambda(convert_image_fn),
+            transforms.Lambda(partial(resize_to_minimum_size, image_size)),
+            transforms.RandomHorizontalFlip(),
+            transforms.Resize(image_size),
+            transforms.CenterCrop(image_size),
+            transforms.ToTensor(),
+            transforms.Lambda(expand_greyscale(num_channels))
+        ])
+
+    def __len__(self):
+        return len(self.paths)
+
+    def __getitem__(self, index):
+        path = self.paths[index]
+        img = Image.open(path)
+        return self.transform(img)
