+++ stylegan2_pytorch/dataset/__init__.py
@@ -1,0 +1,1 @@
+from .Dataset import *
