+++ stylegan2_pytorch/gan/__init__.py
@@ -1,0 +1,1 @@
+from .GAN import *
