+++ stylegan2_pytorch/gan/discriminator/__init__.py
@@ -1,0 +1,1 @@
+from .Discriminator import Discriminator
