+++ stylegan2_pytorch/gan/ema/EMA.py
@@ -1,0 +1,10 @@
+
+
+class EMA:
+    def __init__(self, beta):
+        super().__init__()
+        self.beta = beta
+    def update_average(self, old, new):
+        if old is None:
+            return new
+        return old * self.beta + (1 - self.beta) * new
