+++ stylegan2_pytorch/gan/ema/__init__.py
@@ -1,0 +1,1 @@
+from .EMA import EMA
