+++ stylegan2_pytorch/gan/generator/Generator.py
@@ -1,0 +1,46 @@
+from math import log2
+
+import torch
+from torch import nn
+
+from .GeneratorBlock import GeneratorBlock
+
+
+class Generator(nn.Module):
+    def __init__(self, image_size, latent_dim, network_capacity = 16, transparent = False):
+        super().__init__()
+        self.image_size = image_size
+        self.latent_dim = latent_dim
+        self.num_layers = int(log2(image_size) - 1)
+
+        init_channels = 4 * network_capacity
+        self.initial_block = nn.Parameter(torch.randn((init_channels, 4, 4)))
+        filters = [init_channels] + [network_capacity * (2 ** (i + 1)) for i in range(self.num_layers)][::-1]
+        in_out_pairs = zip(filters[0:-1], filters[1:])
+
+        self.blocks = nn.ModuleList([])
+        for ind, (in_chan, out_chan) in enumerate(in_out_pairs):
+            not_first = ind != 0
+            not_last = ind != (self.num_layers - 1)
+
+            block = GeneratorBlock(
+                latent_dim,
+                in_chan,
+                out_chan,
+                upsample = not_first,
+                upsample_rgb = not_last,
+                rgba = transparent
+            )
+            self.blocks.append(block)
+
+    def forward(self, styles, input_noise):
+        batch_size = styles.shape[0]
+        image_size = self.image_size
+        x = self.initial_block.expand(batch_size, -1, -1, -1)
+        styles = styles.transpose(0, 1)
+
+        rgb = None
+        for style, block in zip(styles, self.blocks):
+            x, rgb = block(x, rgb, style, input_noise)
+
+        return rgb
