+++ stylegan2_pytorch/gan/generator/RGBBlock.py
@@ -1,0 +1,28 @@
+from torch import nn
+
+from .Conv2DMod import Conv2DMod
+
+
+class RGBBlock(nn.Module):
+    def __init__(self, latent_dim, input_channel, upsample, rgba = False):
+        super().__init__()
+        self.input_channel = input_channel
+        self.to_style = nn.Linear(latent_dim, input_channel)
+
+        out_filters = 3 if not rgba else 4
+        self.conv = Conv2DMod(input_channel, out_filters, 1, demod=False)
+
+        self.upsample = nn.Upsample(scale_factor = 2, mode='bilinear', align_corners=False) if upsample else None
+
+    def forward(self, x, prev_rgb, istyle):
+        b, c, h, w = x.shape
+        style = self.to_style(istyle)
+        x = self.conv(x, style)
+
+        if prev_rgb is not None:
+            x = x + prev_rgb
+
+        if self.upsample is not None:
+            x = self.upsample(x)
+
+        return x
