+++ stylegan2_pytorch/gan/generator/__init__.py
@@ -1,0 +1,1 @@
+from .Generator import Generator
