+++ stylegan2_pytorch/gan/style/__init__.py
@@ -1,0 +1,1 @@
+from .StyleVectorizer import StyleVectorizer
