+++ stylegan2_pytorch/misc/constants.py
@@ -1,0 +1,2 @@
+EXTS = ['jpg', 'png']
+EPS = 1e-8
