+++ stylegan2_pytorch/misc/exceptions.py
@@ -1,0 +1,2 @@
+class NanException(Exception):
+    pass
