+++ stylegan2_pytorch/misc/helpers.py
@@ -1,0 +1,98 @@
+import torch
+from torch.autograd import grad as torch_grad
+from torch import nn
+import torchvision
+
+from .exceptions import NanException
+
+try:
+    from apex import amp
+    APEX_AVAILABLE = True
+except:
+    APEX_AVAILABLE = False
+
+# helpers
+
+def default(value, d):
+    return d if value is None else value
+
+def cycle(iterable):
+    while True:
+        for i in iterable:
+            yield i
+
+def raise_if_nan(t):
+    if torch.isnan(t):
+        raise NanException
+
+def loss_backwards(fp16, loss, optimizer, **kwargs):
+    if fp16:
+        with amp.scale_loss(loss, optimizer) as scaled_loss:
+            scaled_loss.backward(**kwargs)
+    else:
+        loss.backward(**kwargs)
+
+def gradient_penalty(images, output, weight = 10):
+    batch_size = images.shape[0]
+    gradients = torch_grad(outputs=output, inputs=images,
+                           grad_outputs=torch.ones(output.size()).cuda(),
+                           create_graph=True, retain_graph=True, only_inputs=True)[0]
+
+    gradients = gradients.view(batch_size, -1)
+    return weight * ((gradients.norm(2, dim=1) - 1) ** 2).mean()
+
+def noise(n, latent_dim):
+    return torch.randn(n, latent_dim).cuda()
+
+def noise_list(n, layers, latent_dim):
+    return [(noise(n, latent_dim), layers)]
+
+def mixed_list(n, layers, latent_dim):
+    tt = int(torch.rand(()).numpy() * layers)
+    return noise_list(n, tt, latent_dim) + noise_list(n, layers - tt, latent_dim)
+
+def latent_to_w(style_vectorizer, latent_descr):
+    return [(style_vectorizer(z), num_layers) for z, num_layers in latent_descr]
+
+def image_noise(n, im_size):
+    return torch.FloatTensor(n, im_size, im_size, 1).uniform_(0., 1.).cuda()
+
+def leaky_relu(p=0.2):
+    return nn.LeakyReLU(p, inplace=True)
+
+def evaluate_in_chunks(max_batch_size, model, *args):
+    split_args = list(zip(*list(map(lambda x: x.split(max_batch_size, dim=0), args))))
+    chunked_outputs = [model(*i) for i in split_args]
+    if len(chunked_outputs) == 1:
+        return chunked_outputs[0]
+    return torch.cat(chunked_outputs, dim=0)
+
+def styles_def_to_tensor(styles_def):
+    return torch.cat([t[:, None, :].expand(-1, n, -1) for t, n in styles_def], dim=1)
+
+def set_requires_grad(model, bool):
+    for p in model.parameters():
+        p.requires_grad = bool
+
+# dataset
+
+def convert_rgb_to_transparent(image):
+    if image.mode == 'RGB':
+        return image.convert('RGBA')
+    return image
+
+def convert_transparent_to_rgb(image):
+    if image.mode == 'RGBA':
+        return image.convert('RGB')
+    return image
+
+class expand_greyscale(object):
+    def __init__(self, num_channels):
+        self.num_channels = num_channels
+    def __call__(self, tensor):
+        return tensor.expand(self.num_channels, -1, -1)
+
+def resize_to_minimum_size(min_size, image):
+    if max(*image.size) < min_size:
+        return torchvision.transforms.functional.resize(image, min_size)
+    return image
